refactor(component-contact): simplify ContactForm
